+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/messages/MessageListViewModel.kt
@@ -74,6 +74,8 @@ import io.getstream.chat.android.offline.extensions.cancelEphemeralMessage
 import io.getstream.chat.android.offline.extensions.getRepliesAsState
 import io.getstream.chat.android.offline.extensions.globalState
 import io.getstream.chat.android.offline.extensions.loadMessageById
+import io.getstream.chat.android.offline.extensions.loadNewerMessages
+import io.getstream.chat.android.offline.extensions.loadNewestMessages
 import io.getstream.chat.android.offline.extensions.loadOlderMessages
 import io.getstream.chat.android.offline.extensions.watchChannelAsState
 import io.getstream.chat.android.offline.plugin.state.channel.ChannelState
@@ -270,7 +272,13 @@ public class MessageListViewModel(
     private fun observeChannel() {
         viewModelScope.launch {
             channelState.filterNotNull().collectLatest { channelState ->
-                combine(channelState.messagesState, user, channelState.reads) { state, user, reads ->
+                combine(
+                    channelState.messagesState,
+                    user,
+                    channelState.reads,
+                    channelState.unreadCount,
+                    channelState.endOfNewerMessages
+                ) { state, user, reads, unreadCount, endOfNewerMessages ->
                     when (state) {
                         is io.getstream.chat.android.offline.plugin.state.channel.MessagesState.NoQueryActive,
                         is io.getstream.chat.android.offline.plugin.state.channel.MessagesState.Loading,
@@ -290,7 +298,11 @@ public class MessageListViewModel(
                                 ),
                                 isLoadingMore = false,
                                 endOfMessages = channelState.endOfOlderMessages.value,
+                                startOfMessages = endOfNewerMessages,
                                 currentUser = user,
+                                isLoadingMoreNewMessages = false,
+                                isLoadingMoreOldMessages = false,
+                                unreadCount = unreadCount ?: messagesState.unreadCount
                             )
                         }
                     }
@@ -312,16 +324,15 @@ public class MessageListViewModel(
 
                             newState.copy(
                                 newMessageState = newMessageState,
-                                unreadCount = getUnreadMessageCount(newMessageState)
                             )
                         } else {
                             newState
                         }
 
-                        messagesState.messageItems.firstOrNull {
-                            it is MessageItemState && it.message.id == scrollToMessage?.id
-                        }?.let {
-                            focusMessage((it as MessageItemState).message.id)
+                        if (scrollToMessage != null) {
+                            messagesState.messageItems.firstOrNull {
+                                it is MessageItemState && it.message.id == scrollToMessage?.id
+                            }?.let { focusMessage((it as MessageItemState).message.id) }
                         }
 
                         lastLoadedMessage = newLastMessage
@@ -505,6 +516,14 @@ public class MessageListViewModel(
     /**
      * Triggered when the user loads more data by reaching the end of the current messages.
      */
+    @Deprecated(
+        message = "Deprecated after implementing bi directional pagination.",
+        replaceWith = ReplaceWith(
+            "loadOlderMessages(messageId: String)",
+            "io.getstream.chat.android.compose.viewmodel.messages"
+        ),
+        level = DeprecationLevel.WARNING
+    )
     public fun loadMore() {
         if (chatClient.clientState.isOffline) return
         val messageMode = messageMode
@@ -512,7 +531,38 @@ public class MessageListViewModel(
         if (messageMode is MessageMode.MessageThread) {
             threadLoadMore(messageMode)
         } else {
-            messagesState = messagesState.copy(isLoadingMore = true)
+            messagesState = messagesState.copy(isLoadingMore = true, isLoadingMoreOldMessages = true)
+            chatClient.loadOlderMessages(channelId, messageLimit).enqueue()
+        }
+    }
+
+    /**
+     * Loads newer messages of a channel following the currently newest loaded message. In case of threads this will
+     * do nothing.
+     *
+     * @param messageId The id of the newest [Message] inside the messages list.
+     */
+    public fun loadNewerMessages(messageId: String) {
+        if (chatClient.globalState.isOffline() || messagesState.startOfMessages) return
+
+        if (messageMode is MessageMode.Normal) {
+            messagesState = messagesState.copy(isLoadingMore = true, isLoadingMoreNewMessages = true)
+            chatClient.loadNewerMessages(channelId, messageId, messageLimit).enqueue()
+        }
+    }
+
+    /**
+     * Loads older messages of a channel following the currently oldest loaded message. Also will load older messages
+     * of a thread.
+     */
+    public fun loadOlderMessages() {
+        if (chatClient.globalState.isOffline() || messagesState.endOfMessages) return
+        val messageMode = messageMode
+
+        if (messageMode is MessageMode.MessageThread) {
+            threadLoadMore(messageMode)
+        } else {
+            messagesState = messagesState.copy(isLoadingMore = true, isLoadingMoreOldMessages = true)
             chatClient.loadOlderMessages(channelId, messageLimit).enqueue()
         }
     }
@@ -542,10 +592,7 @@ public class MessageListViewModel(
      * @param message The selected message we wish to scroll to.
      */
     private fun loadMessage(message: Message) {
-        val cid = channelState.value?.cid
-        if (cid == null || chatClient.clientState.isOffline) return
-
-        chatClient.loadMessageById(cid, message.id).enqueue()
+        chatClient.loadMessageById(channelId, message.id).enqueue()
     }
 
     /**
@@ -723,7 +770,9 @@ public class MessageListViewModel(
                     isLoadingMore = false,
                     endOfMessages = endOfOlderMessages,
                     currentUser = user,
-                    parentMessageId = threadId
+                    parentMessageId = threadId,
+                    isLoadingMoreOldMessages = false,
+                    isLoadingMoreNewMessages = false
                 )
             }.collect { newState -> threadMessagesState = newState }
         }
@@ -968,7 +1017,9 @@ public class MessageListViewModel(
      * or "New Message" actions in the list or simply scrolls to the bottom.
      */
     public fun clearNewMessageState() {
+        if (!messagesState.startOfMessages) return
         threadMessagesState = threadMessagesState.copy(newMessageState = null, unreadCount = 0)
+
         messagesState = messagesState.copy(newMessageState = null, unreadCount = 0)
     }
 
@@ -1019,12 +1070,14 @@ public class MessageListViewModel(
      * Updates the current message state with new messages.
      *
      * @param messages The list of new message items.
-     */
+     * */
     private fun updateMessages(messages: List<MessageListItemState>) {
         if (isInThread) {
-            this.threadMessagesState = threadMessagesState.copy(messageItems = messages)
+            this.threadMessagesState =
+                threadMessagesState.copy(messageItems = messages)
         } else {
-            this.messagesState = messagesState.copy(messageItems = messages)
+            this.messagesState =
+                messagesState.copy(messageItems = messages)
         }
     }
 
@@ -1056,20 +1109,47 @@ public class MessageListViewModel(
     }
 
     /**
-     * Scrolls to message if in list otherwise get the message from backend.
+     * Scrolls to message if in list otherwise get the message from backend. Does not work for threads.
      *
      * @param message The message we wish to scroll to.
      */
     public fun scrollToSelectedMessage(message: Message) {
+        if (isInThread) return
+
         val isMessageInList = currentMessagesState.messageItems.firstOrNull {
             it is MessageItemState && it.message.id == message.id
         } != null
+        scrollToMessage = message
 
         if (isMessageInList) {
             focusMessage(message.id)
         } else {
-            scrollToMessage = message
             loadMessage(message = message)
+        }
+    }
+
+    /**
+     * Requests that the list scrolls to the bottom to the newest messages. If the newest messages are loaded will set
+     * scroll the list to the bottom. If they are not loaded will request the newest data and once loaded will scroll
+     * to the bottom of the list.
+     *
+     * @param messageLimit The message count we wish to load from the API when loading new messages.
+     * @param scrollToBottom Notifies the ui to scroll to the bottom if the newest messages are in the list or have been
+     * loaded from the API.
+     */
+    public fun scrollToBottom(messageLimit: Int = DefaultMessageLimit, scrollToBottom: () -> Unit) {
+        if (isInThread) {
+            scrollToBottom()
+        } else {
+            if (currentMessagesState.startOfMessages) {
+                scrollToBottom()
+                return
+            }
+            chatClient.loadNewestMessages(channelId, messageLimit).enqueue {
+                if (it.isSuccess) {
+                    scrollToBottom()
+                }
+            }
         }
     }