+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/extensions/ChatClient.kt
@@ -383,3 +383,34 @@ private suspend fun ChatClient.loadMessageByIdInternal(
         Result(ChatError("Error while fetching messages from backend. Messages around id: $messageId"))
     }
 }
+
+/**
+ * Loads the newest messages of a channel.
+ *
+ * @param cid The full channel id i. e. messaging:123.
+ * @param messageLimit The number of messages to be loaded.
+ * @param userPresence Flag to determine if the SDK is going to receive UserPresenceChanged events.
+ * Used by the SDK to indicate if the user is online or not.
+ *
+ * @return Executable async [Call] responsible for loading the newest messages.
+ */
+public fun ChatClient.loadNewestMessages(cid: String, messageLimit: Int, userPresence: Boolean = true): Call<Channel> {
+    return CoroutineCall(state.scope) {
+        loadNewestMessagesInternal(cid, messageLimit, userPresence)
+    }
+}
+
+@CheckResult
+internal suspend fun ChatClient.loadNewestMessagesInternal(
+    cid: String,
+    messageLimit: Int,
+    userPresence: Boolean
+): Result<Channel> {
+    val cidValidationResult = validateCidWithResult(cid)
+    if (!cidValidationResult.isSuccess) {
+        return cidValidationResult.error().toResultError()
+    }
+
+    val (channelType, channelId) = cid.cidToTypeAndId()
+    return logic.channel(channelType = channelType, channelId = channelId).watch(messageLimit, userPresence)
+}