+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/logic/channel/internal/ChannelStateLogicImpl.kt
@@ -242,18 +242,20 @@ internal class ChannelStateLogicImpl(
     }
 
     /**
-     * Upsert members in the channel.
+     * Upsert messages in the channel.
      *
      * @param message The message to be added or updated.
      */
     override fun upsertMessage(message: Message) {
-        upsertMessages(listOf(message))
+        if (mutableState.rawMessages.containsKey(message.id) || mutableState.endOfNewerMessages.value) {
+            upsertMessages(listOf(message))
+        }
     }
 
     /**
-     * Upsert members in the channel.
+     * Upsert messages in the channel.
      *
-     * @param messages the list of [Message] to be upserted
+     * @param messages the list of [Message] to be upserted.
      * @param shouldRefreshMessages if the current messages should be removed or not and only
      * new messages should be kept.
      */
@@ -460,19 +462,37 @@ internal class ChannelStateLogicImpl(
         searchLogic.handleMessageBounds(request, noMoreMessages)
         mutableState.recoveryNeeded = false
 
-        if (noMoreMessages) {
-            if (request.isFilteringNewerMessages()) {
+        determinePaginationEnd(request, noMoreMessages)
+
+        updateDataFromChannel(
+            channel,
+            shouldRefreshMessages = request.isFilteringAroundIdMessages() ||
+                (!request.isFilteringMessages() && !mutableState.insideSearch.value),
+            scrollUpdate = request.isFilteringNewerMessages() || request.filteringOlderMessages()
+        )
+    }
+
+    private fun determinePaginationEnd(request: QueryChannelRequest, noMoreMessages: Boolean) {
+        when {
+            /* If we are not filtering the messages in any direction and not providing any message id then
+            * we are requesting the newest messages, only if not inside search so we don't override the
+            * search results */
+            !request.isFilteringMessages() -> {
+                mutableState.setEndOfOlderMessages(false)
+                mutableState.setEndOfNewerMessages(!mutableState.insideSearch.value)
+            }
+            /* If we are filtering around a specific message we are loading both newer and older messages
+            * and can't be sure if there are no older or newer messages left */
+            request.isFilteringAroundIdMessages() -> {
+                mutableState.setEndOfOlderMessages(false)
+                mutableState.setEndOfNewerMessages(false)
+            }
+            noMoreMessages -> if (request.isFilteringNewerMessages()) {
                 mutableState.setEndOfNewerMessages(true)
             } else {
                 mutableState.setEndOfOlderMessages(true)
             }
         }
-
-        updateDataFromChannel(
-            channel,
-            shouldRefreshMessages = request.isFilteringAroundIdMessages(),
-            scrollUpdate = request.isFilteringNewerMessages() || request.filteringOlderMessages()
-        )
     }
 
     /**