+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/internal/MessageListScrollHelper.kt
@@ -26,6 +26,7 @@ import io.getstream.chat.android.ui.common.extensions.internal.dpToPx
 import io.getstream.chat.android.ui.common.extensions.internal.getFragmentManager
 import io.getstream.chat.android.ui.common.extensions.internal.safeCast
 import io.getstream.chat.android.ui.common.extensions.isDeleted
+import io.getstream.chat.android.ui.message.list.MessageListView
 import io.getstream.chat.android.ui.message.list.adapter.BaseMessageItemViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.internal.MessageListItemAdapter
 import kotlin.math.max
@@ -47,6 +48,11 @@ internal class MessageListScrollHelper(
 
     private var lastSeenMessageInChannel: MessageListItem? = null
     private var lastSeenMessageInThread: MessageListItem? = null
+
+    private var onScrollToBottomHandler: MessageListView.OnScrollToBottomHandler =
+        MessageListView.OnScrollToBottomHandler {
+            recyclerView.scrollToPosition(currentList.lastIndex)
+        }
 
     internal var unreadCountEnabled: Boolean = true
 
@@ -73,7 +79,7 @@ internal class MessageListScrollHelper(
 
     init {
         scrollButtonView.setOnClickListener {
-            recyclerView.scrollToPosition(currentList.lastIndex)
+            onScrollToBottomHandler.onScrollToBottom()
         }
         recyclerView.addOnScrollListener(
             object : RecyclerView.OnScrollListener() {
@@ -168,14 +174,23 @@ internal class MessageListScrollHelper(
         )
     }
 
-    internal fun onMessageListChanged(isThreadStart: Boolean, hasNewMessages: Boolean, isInitialList: Boolean) {
-        if (!scrollToBottomButtonEnabled || (!hasNewMessages || adapter.currentList.isEmpty())) {
+    internal fun scrollToBottom() {
+        recyclerView.scrollToPosition(currentList.lastIndex)
+    }
+
+    internal fun onMessageListChanged(
+        isThreadStart: Boolean,
+        hasNewMessages: Boolean,
+        isInitialList: Boolean,
+        areNewestMessagesLoaded: Boolean,
+    ) {
+        if (shouldKeepScrollPosition(areNewestMessagesLoaded, hasNewMessages)) {
             return
         }
 
         if (isThreadStart) {
             layoutManager.scrollToPosition(0)
-        } else if (isInitialList || isLastMessageMine() || isAtBottom || alwaysScrollToBottom) {
+        } else if (shouldScrollToBottom(isInitialList, areNewestMessagesLoaded)) {
             layoutManager.scrollToPosition(currentList.lastIndex)
             callback.onLastMessageRead()
         } else {
@@ -183,6 +198,21 @@ internal class MessageListScrollHelper(
             scrollButtonView.setUnreadCount(unreadCount)
             scrollButtonView.isVisible = true
         }
+    }
+
+    private fun shouldKeepScrollPosition(
+        areNewestMessagesLoaded: Boolean,
+        hasNewMessages: Boolean,
+    ): Boolean {
+        return areNewestMessagesLoaded || !scrollToBottomButtonEnabled ||
+            (!hasNewMessages || adapter.currentList.isEmpty())
+    }
+
+    private fun shouldScrollToBottom(
+        isInitialList: Boolean,
+        areNewestMessagesLoaded: Boolean,
+    ): Boolean {
+        return areNewestMessagesLoaded && (isInitialList || isLastMessageMine() || isAtBottom || alwaysScrollToBottom)
     }
 
     private fun refreshUnreadCount() {
@@ -216,6 +246,10 @@ internal class MessageListScrollHelper(
 
     private fun MessageListItem.isValid(): Boolean {
         return this is MessageListItem.MessageItem && !(this.isTheirs && this.message.isDeleted())
+    }
+
+    internal fun setScrollToBottomHandler(onScrollToBottomHandler: MessageListView.OnScrollToBottomHandler) {
+        this.onScrollToBottomHandler = onScrollToBottomHandler
     }
 
     internal fun interface MessageReadListener {